+++ py/desimodel/inputs/throughput.py
@@ -1,5 +1,10 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+# -*- coding: utf-8 -*-
 '''
-Utilities for updating throughput model
+desimodel.inputs.throughput
+===========================
+
+Utilities for updating throughput model.
 '''
 import os
 import shutil